requirements.txt
slack_bolt
google-genai
mdutils
fantasynames
anthropic
requests
python-dotenv
prompt_toolkit
discord.py
